requirements.txt
streamlit
openai
faiss-cpu
tiktoken
python-dotenv
requests
gdown

